Fix failing public tests
Package click==8.2.0 removed mix_stderr parameter from CliRunner. This
is only used in tests and not in the actual code. The fix is not
trivial. Therefore, temporary hardcode a lower click version for tests
to make them pass.

IPython is not parsing misindented magic methods anymore. Therefore,
simplify a notebook used to test formatting.

PiperOrigin-RevId: 763795143